Refactor functions into separate file: threeplay_sync.py
